requirements.txt
discord.py
colorama
requests
aiohttp
tengiphpy
nekos.py
psutil
wikipedia

